+++ RealmSwift-swift2.0/List.swift
@@ -51,16 +51,15 @@ When added as a property on `Object` models, the property must be declared as `l
 */
 public final class List<T: Object>: ListBase {
 
+    /// Element type contained in this collection.
+    public typealias Element = T
+
     // MARK: Properties
 
-    /// The Realm the objects in this list belong to, or `nil` if the list's owning
-    /// object does not belong to a realm (the list is standalone).
+    /// The Realm the objects in this list belong to, or `nil` if the list's
+    /// owning object does not belong to a realm (the list is standalone).
     public var realm: Realm? {
-        if let rlmRealm = _rlmArray.realm {
-            return Realm(rlmRealm)
-        } else {
-            return nil
-        }
+        return _rlmArray.realm.map { Realm($0) }
     }
 
     /// Indicates if the list can no longer be accessed.
@@ -215,6 +214,60 @@ public final class List<T: Object>: ListBase {
         return Results<T>(_rlmArray.sortedResultsUsingDescriptors(sortDescriptors.map { $0.rlmSortDescriptorValue }))
     }
 
+    // MARK: Aggregate Operations
+
+    /**
+    Returns the minimum value of the given property.
+
+    - warning: Only names of properties of a type conforming to the `MinMaxType` protocol can be used.
+
+    - parameter property: The name of a property conforming to `MinMaxType` to look for a minimum on.
+
+    - returns: The minimum value for the property amongst objects in the List, or `nil` if the List is empty.
+    */
+    public func min<U: MinMaxType>(property: String) -> U? {
+        return filter(NSPredicate(value: true)).min(property)
+    }
+
+    /**
+    Returns the maximum value of the given property.
+
+    - warning: Only names of properties of a type conforming to the `MinMaxType` protocol can be used.
+
+    - parameter property: The name of a property conforming to `MinMaxType` to look for a maximum on.
+
+    - returns: The maximum value for the property amongst objects in the List, or `nil` if the List is empty.
+    */
+    public func max<U: MinMaxType>(property: String) -> U? {
+        return filter(NSPredicate(value: true)).max(property)
+    }
+
+    /**
+    Returns the sum of the given property for objects in the List.
+
+    - warning: Only names of properties of a type conforming to the `AddableType` protocol can be used.
+
+    - parameter property: The name of a property conforming to `AddableType` to calculate sum on.
+
+    - returns: The sum of the given property over all objects in the List.
+    */
+    public func sum<U: AddableType>(property: String) -> U {
+        return filter(NSPredicate(value: true)).sum(property)
+    }
+
+    /**
+    Returns the average of the given property for objects in the List.
+
+    - warning: Only names of properties of a type conforming to the `AddableType` protocol can be used.
+
+    - parameter property: The name of a property conforming to `AddableType` to calculate average on.
+
+    - returns: The average of the given property over all objects in the List, or `nil` if the List is empty.
+    */
+    public func average<U: AddableType>(property: String) -> U? {
+        return filter(NSPredicate(value: true)).average(property)
+    }
+
     // MARK: Mutation
 
     /**
@@ -336,7 +389,7 @@ public final class List<T: Object>: ListBase {
     }
 }
 
-extension List: RangeReplaceableCollectionType {
+extension List: RealmCollectionType, RangeReplaceableCollectionType {
     // MARK: Sequence Support
 
     /// Returns a `GeneratorOf<T>` that yields successive elements in the list.